AG-10217 - Build the same default format sting for both timeScale ticks and ordinalTimeScale ticks
