+++ packages/ag-charts-community/src/scale/ordinalTimeScale.ts
@@ -99,83 +99,7 @@ export class OrdinalTimeScale extends BandScale<Date> {
      * If no specifier is provided, this method returns the default time format function.
      */
     tickFormat({ ticks, specifier }: { ticks?: any[]; specifier?: string }): (date: Date) => string {
-        return specifier == undefined
-            ? defaultTimeTickFormat(this.buildFormatString, ticks)
-            : buildFormatter(specifier);
-    }
-
-    buildFormatString(defaultTimeFormat: DefaultTimeFormats, yearChange: boolean, ticks: any[]): string {
-        let formatStringArray: string[] = [];
-        let timeEndIndex = 0;
-
-        const firstTick = ticks[0];
-        const secondTick = ticks[1];
-
-        if (ticks.length === 0 || !(firstTick instanceof Date) || !(secondTick instanceof Date)) {
-            return ``;
-        }
-
-        switch (defaultTimeFormat) {
-            case DefaultTimeFormats.SECOND:
-                if (Math.abs(firstTick.getSeconds() - secondTick.getSeconds()) > 0) {
-                    formatStringArray.push(TIME_FORMAT_STRINGS[DefaultTimeFormats.MINUTE]);
-                }
-            // fall through deliberately
-            case DefaultTimeFormats.MINUTE:
-                if (Math.abs(firstTick.getMinutes() - secondTick.getMinutes()) > 0) {
-                    formatStringArray.push(TIME_FORMAT_STRINGS[DefaultTimeFormats.HOUR]);
-                }
-            // fall through deliberately
-            case DefaultTimeFormats.HOUR:
-                if (Math.abs(firstTick.getHours() - secondTick.getHours()) > 0) {
-                    formatStringArray.push(TIME_FORMAT_STRINGS[DefaultTimeFormats.WEEK_DAY]);
-                }
-            // fall through deliberately
-            case DefaultTimeFormats.WEEK_DAY:
-                timeEndIndex = formatStringArray.length;
-                if (Math.abs(firstTick.getMonth() - secondTick.getMonth()) > 0 || yearChange) {
-                    formatStringArray.push(TIME_FORMAT_STRINGS[DefaultTimeFormats.SHORT_MONTH]);
-                } else if (Math.abs(firstTick.getDay() - secondTick.getDay()) > 0) {
-                    formatStringArray.push(TIME_FORMAT_STRINGS[DefaultTimeFormats.WEEK_DAY]);
-                }
-            // fall through deliberately
-            case DefaultTimeFormats.SHORT_MONTH:
-            case DefaultTimeFormats.MONTH:
-                const monthIndex = formatStringArray.indexOf(TIME_FORMAT_STRINGS[DefaultTimeFormats.SHORT_MONTH]);
-
-                if (monthIndex < 0) {
-                    formatStringArray.push(TIME_FORMAT_STRINGS[DefaultTimeFormats.SHORT_MONTH]);
-                }
-            // fall through deliberately
-            case DefaultTimeFormats.YEAR:
-            default:
-                if (Math.abs(firstTick.getFullYear() - secondTick.getFullYear()) > 0 || yearChange) {
-                    formatStringArray.push(TIME_FORMAT_STRINGS[DefaultTimeFormats.YEAR]);
-                }
-                break;
-        }
-
-        if (timeEndIndex < formatStringArray.length) {
-            // Insert a gap between all date components.
-            formatStringArray = [
-                ...formatStringArray.slice(0, timeEndIndex),
-                formatStringArray.slice(timeEndIndex).join(' '),
-            ];
-        }
-        if (timeEndIndex > 0) {
-            // Reverse order of time components, since they should be displayed in descending
-            // granularity.
-            formatStringArray = [
-                ...formatStringArray.slice(0, timeEndIndex).reverse(),
-                ...formatStringArray.slice(timeEndIndex),
-            ];
-            if (timeEndIndex < formatStringArray.length) {
-                // Insert a gap between time and date components.
-                formatStringArray.splice(timeEndIndex, 0, ' ');
-            }
-        }
-
-        return formatStringArray.join('');
+        return specifier == undefined ? defaultTimeTickFormat(ticks) : buildFormatter(specifier);
     }
 
     override invert(y: number): Date {
