+++ packages/ag-charts-community/src/scale/timeScale.ts
@@ -82,74 +82,6 @@ export class TimeScale extends ContinuousScale<Date, TimeInterval | number> {
         return new Date(d);
     }
 
-    buildFormatString(defaultTimeFormat: DefaultTimeFormats, yearChange: boolean, ticks: any[]): string {
-        let formatStringArray: string[] = [TIME_FORMAT_STRINGS[defaultTimeFormat]];
-        let timeEndIndex = 0;
-
-        const firstTick = dateToNumber(ticks[0]);
-        const lastTick = dateToNumber(ticks.at(-1)!);
-        const extent = Math.abs(lastTick - firstTick);
-
-        switch (defaultTimeFormat) {
-            case DefaultTimeFormats.SECOND:
-                if (extent / durationMinute > 1) {
-                    formatStringArray.push(TIME_FORMAT_STRINGS[DefaultTimeFormats.MINUTE]);
-                }
-            // fall through deliberately
-            case DefaultTimeFormats.MINUTE:
-                if (extent / durationHour > 1) {
-                    formatStringArray.push(TIME_FORMAT_STRINGS[DefaultTimeFormats.HOUR]);
-                }
-            // fall through deliberately
-            case DefaultTimeFormats.HOUR:
-                timeEndIndex = formatStringArray.length;
-                if (extent / durationDay > 1) {
-                    formatStringArray.push(TIME_FORMAT_STRINGS[DefaultTimeFormats.WEEK_DAY]);
-                }
-            // fall through deliberately
-            case DefaultTimeFormats.WEEK_DAY:
-                if (extent / durationWeek > 1 || yearChange) {
-                    // if it's more than a week or there is a year change, don't show week day
-                    const weekDayIndex = formatStringArray.indexOf(TIME_FORMAT_STRINGS[DefaultTimeFormats.WEEK_DAY]);
-
-                    if (weekDayIndex > -1) {
-                        formatStringArray.splice(weekDayIndex, 1, TIME_FORMAT_STRINGS[DefaultTimeFormats.SHORT_MONTH]);
-                    }
-                }
-            // fall through deliberately
-            case DefaultTimeFormats.SHORT_MONTH:
-            case DefaultTimeFormats.MONTH:
-                if (extent / durationYear > 1 || yearChange) {
-                    formatStringArray.push(TIME_FORMAT_STRINGS[DefaultTimeFormats.YEAR]);
-                }
-            // fall through deliberately
-            default:
-                break;
-        }
-
-        if (timeEndIndex < formatStringArray.length) {
-            // Insert a gap between all date components.
-            formatStringArray = [
-                ...formatStringArray.slice(0, timeEndIndex),
-                formatStringArray.slice(timeEndIndex).join(' '),
-            ];
-        }
-        if (timeEndIndex > 0) {
-            // Reverse order of time components, since they should be displayed in descending
-            // granularity.
-            formatStringArray = [
-                ...formatStringArray.slice(0, timeEndIndex).reverse(),
-                ...formatStringArray.slice(timeEndIndex),
-            ];
-            if (timeEndIndex < formatStringArray.length) {
-                // Insert a gap between time and date components.
-                formatStringArray.splice(timeEndIndex, 0, ' ');
-            }
-        }
-
-        return formatStringArray.join('');
-    }
-
     /**
      * @param options Tick interval options.
      * @param options.start The start time (timestamp).
@@ -296,9 +228,7 @@ export class TimeScale extends ContinuousScale<Date, TimeInterval | number> {
      * If no specifier is provided, this method returns the default time format function.
      */
     tickFormat({ ticks, specifier }: { ticks?: any[]; specifier?: string }): (date: Date) => string {
-        return specifier == undefined
-            ? defaultTimeTickFormat(this.buildFormatString, ticks)
-            : buildFormatter(specifier);
+        return specifier == undefined ? defaultTimeTickFormat(ticks) : buildFormatter(specifier);
     }
 
     update() {
