+++ packages/ag-charts-community/src/util/timeFormatDefaults.ts
@@ -5,6 +5,7 @@ import timeMonth from '../util/time/month';
 import timeSecond from '../util/time/second';
 import timeWeek from '../util/time/week';
 import timeYear from '../util/time/year';
+import { durationDay, durationHour, durationMinute, durationWeek, durationYear } from './time/duration';
 import { buildFormatter } from './timeFormat';
 
 export enum DefaultTimeFormats {
@@ -31,16 +32,16 @@ export const TIME_FORMAT_STRINGS: Record<DefaultTimeFormats, string> = {
     [DefaultTimeFormats.YEAR]: '%Y',
 };
 
-type FormatStringBuilderFn = (defaultTimeFormat: DefaultTimeFormats, yearChange: boolean, ticks: any[]) => string;
-
 export function dateToNumber(x: any) {
     return x instanceof Date ? x.getTime() : x;
 }
 
-export function calculateDefaultTimeTickFormat(
-    ticks: any[] | undefined = [],
-    formatStringBuilder: FormatStringBuilderFn
-) {
+export function defaultTimeTickFormat(ticks?: any[]) {
+    const formatString = calculateDefaultTimeTickFormat(ticks);
+    return (date: Date) => buildFormatter(formatString)(date);
+}
+
+export function calculateDefaultTimeTickFormat(ticks: any[] | undefined = []) {
     let defaultTimeFormat = DefaultTimeFormats.YEAR as DefaultTimeFormats;
 
     const updateFormat = (format: DefaultTimeFormats) => {
@@ -84,7 +85,70 @@ export function getLowestGranularityFormat(value: Date | number): DefaultTimeFor
     return DefaultTimeFormats.YEAR;
 }
 
-export function defaultTimeTickFormat(formatStringBuilder: FormatStringBuilderFn, ticks?: any[]) {
-    const formatString = calculateDefaultTimeTickFormat(ticks, formatStringBuilder);
-    return (date: Date) => buildFormatter(formatString)(date);
+export function formatStringBuilder(defaultTimeFormat: DefaultTimeFormats, yearChange: boolean, ticks: any[]): string {
+    let formatStringArray: string[] = [TIME_FORMAT_STRINGS[defaultTimeFormat]];
+    let timeEndIndex = 0;
+
+    const firstTick = dateToNumber(ticks[0]);
+    const lastTick = dateToNumber(ticks.at(-1)!);
+    const extent = Math.abs(lastTick - firstTick);
+
+    switch (defaultTimeFormat) {
+        case DefaultTimeFormats.SECOND:
+            if (extent / durationMinute > 1) {
+                formatStringArray.push(TIME_FORMAT_STRINGS[DefaultTimeFormats.MINUTE]);
+            }
+        // fall through deliberately
+        case DefaultTimeFormats.MINUTE:
+            if (extent / durationHour > 1) {
+                formatStringArray.push(TIME_FORMAT_STRINGS[DefaultTimeFormats.HOUR]);
+            }
+        // fall through deliberately
+        case DefaultTimeFormats.HOUR:
+            timeEndIndex = formatStringArray.length;
+            if (extent / durationDay > 1) {
+                formatStringArray.push(TIME_FORMAT_STRINGS[DefaultTimeFormats.WEEK_DAY]);
+            }
+        // fall through deliberately
+        case DefaultTimeFormats.WEEK_DAY:
+            if (extent / durationWeek > 1 || yearChange) {
+                // if it's more than a week or there is a year change, don't show week day
+                const weekDayIndex = formatStringArray.indexOf(TIME_FORMAT_STRINGS[DefaultTimeFormats.WEEK_DAY]);
+
+                if (weekDayIndex > -1) {
+                    formatStringArray.splice(weekDayIndex, 1, TIME_FORMAT_STRINGS[DefaultTimeFormats.SHORT_MONTH]);
+                }
+            }
+        // fall through deliberately
+        case DefaultTimeFormats.SHORT_MONTH:
+        case DefaultTimeFormats.MONTH:
+            if (extent / durationYear > 1 || yearChange) {
+                formatStringArray.push(TIME_FORMAT_STRINGS[DefaultTimeFormats.YEAR]);
+            }
+        // fall through deliberately
+        default:
+            break;
+    }
+
+    if (timeEndIndex < formatStringArray.length) {
+        // Insert a gap between all date components.
+        formatStringArray = [
+            ...formatStringArray.slice(0, timeEndIndex),
+            formatStringArray.slice(timeEndIndex).join(' '),
+        ];
+    }
+    if (timeEndIndex > 0) {
+        // Reverse order of time components, since they should be displayed in descending
+        // granularity.
+        formatStringArray = [
+            ...formatStringArray.slice(0, timeEndIndex).reverse(),
+            ...formatStringArray.slice(timeEndIndex),
+        ];
+        if (timeEndIndex < formatStringArray.length) {
+            // Insert a gap between time and date components.
+            formatStringArray.splice(timeEndIndex, 0, ' ');
+        }
+    }
+
+    return formatStringArray.join('');
 }
